transfered the code
transfered the test funcs from main into new module in 'test' packet
